Refactor formatter implementation
