+++ packages/bento-design-system/src/Charts/ValueFormatter.ts
@@ -1,0 +1,4 @@
+export type NameType = number | string;
+export type ValueType = number | string;
+
+export type ValueFormatter = (value: ValueType) => string;
