+++ packages/bento-design-system/src/Charts/index.ts
@@ -3,3 +3,4 @@ export * from "./ChartProps";
 export * from "./Config";
 export * from "./DonutChart/DonutChart";
 export * from "./LineChart/LineChart";
+export * from "./ValueFormatter";
