--- packages/bento-design-system/src/Charts/utils.ts
@@ -1,6 +1,0 @@
-export type ValueType = number | string;
-export type NameType = number | string;
-
-export type ValueFormatter = (value: ValueType) => string;
-
-export const defaultValueFormatter: ValueFormatter = (value: ValueType) => value.toString();
